src/main/java/com/epam/TravelBooking/service/impl/HotelServiceImpl.java
package com.epam.TravelBooking.service.impl;

import com.epam.TravelBooking.model.Hotel;
import com.epam.TravelBooking.repository.HotelRepository;
import com.epam.TravelBooking.service.HotelService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class HotelServiceImpl implements HotelService {

    private final HotelRepository hotelRepository;

    @Autowired
    public HotelServiceImpl(HotelRepository hotelRepository) {
        this.hotelRepository = hotelRepository;
    }

    @Override
    public List<Hotel> getAllHotels() {
        return hotelRepository.findAll();
    }

    @Override
    public Hotel getHotelById(Long id) {
        return hotelRepository.findById(id).orElse(null);
    }

    @Override
    public Hotel saveHotel(Hotel hotel) {
        return hotelRepository.save(hotel);
    }

    @Override
    public void deleteHotel(Long id) {
        hotelRepository.deleteById(id);
    }

    @Override
    public List<Hotel> searchHotels(String location) {
        return hotelRepository.findByLocation(location);
    }

    // Implement other methods from HotelService interface
}

